Convert custom_app to be typesafe
